### 2021-11-12 #3911 @kyleniemeyer
@jmschrei ok, I will see if I can find someone else. But, I will say that it's usually less important for the editor to be an expert on the language (vs. the domain) and much more important that the reviewers are.
