### 2021-11-16 #3911 @nhejazi
Thanks for editing this submission @mikldk. The ideal reviewer would have a working knowledge of R and some familiarity with causal inference and machine learning. Perhaps someone (or a few) among @jjharden, @csinva, @tedinburgh, @mariacuellar, and @tgerke might have time to review this?
