### 2021-12-03 #3911 @mikldk
> Hi all, I'm sorry for the delay in response. I'm not used to getting article reviews through Github and thought the initial request was spam. What is the timeline for the review? I have a lot going on through the holidays (I'm an AE for an applied stats journal and have a few manuscripts on my plate already).

Thanks, I'll keep you in mind. I would prefer to finished this off before the holidays.
