### 2021-12-06 #3911 @rrrlw
@mikldk I have no expertise in the topic and, thus, would not be useful in reviewing that aspect of the submission. However, I would be willing to review the R package itself (documentation, etc.) if another reviewer who is an expert in this area would be willing to evaluate the functionality/validity.

I can start the review as early as 13-Dec-2021 (1 week from today). Let me know if this works, happy to help if it does!
