### 2021-12-06 #3911 @nhejazi
Just to add a bit of background, @rrrlw and @mikldk, the statistical methodology need not be the subject of the review (though comments are welcome), since this package is an implementation that accompanied work that has already been peer reviewed and published (https://doi.org/10.1093/biomet/asaa085). The primary purpose of this submission is a review of the software.
