### 2021-12-07 #3979 @whedon
[REVIEW]: medoutcon: Nonparametric efficient causal mediation analysis with machine learning in R
**Submitting author:** @nhejazi (<a href="http://orcid.org/0000-0002-7127-2789">Nima Hejazi</a>)
**Repository:** <a href="https://github.com/nhejazi/medoutcon" target ="_blank">https://github.com/nhejazi/medoutcon</a>
**Version:** v0.1.6
