### 2021-12-07 #3979 @mikldk
> Failed to discover a `Statement of need` section in paper

@nhejazi, I think it is fine to rename "Background and Statement of Need" to just "Statement of Need" for consistency purposes.
