### 2021-12-07 #3979 @mikldk
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1093/biomet/asaa085 is OK
> - 10.1111/rssb.12362 is OK
> - 10.5281/zenodo.1342293 is OK
> - 10.1111/ectj.12097 is OK
> - 10.1515/jci-2020-0018 is OK
> - 10.1214/12-AOS990 is OK
> - 10.2202/1557-4679.1361 is OK
> - 10.21105/joss.00512 is OK
> - 10.5281/zenodo.835602 is OK
> - 10.1111/biom.13375 is OK
> - 10.1111/add.15377 is OK
>
> MISSING DOIs
>
> - 10.1007/978-3-319-65304-4_14 may be a valid DOI for title: Stochastic Treatment Regimes
> - 10.1111/j.1541-0420.2011.01685.x may be a valid DOI for title: Population intervention causal effects based on stochastic interventions
> - 10.1515/jci-2016-0006 may be a valid DOI for title: Longitudinal mediation analysis with time-varying mediators and exposures, with application to survival outcomes
> - 10.1097/ede.0000000000000034 may be a valid DOI for title: Effect decomposition in the presence of an exposure-induced mediator-outcome confounder
> - 10.1515/em-2017-0007 may be a valid DOI for title: Robust and flexible estimation of stochastic mediation effects: a proposed method and example in a randomized trial setting
> - 10.1007/978-1-4612-5769-1 may be a valid DOI for title: Contributions to a general asymptotic statistical theory
> - 10.1007/978-1-4419-9782-1_27 may be a valid DOI for title: Cross-validated targeted minimum-loss-based estimation
> - 10.1214/aos/1176350609 may be a valid DOI for title: Consistent estimation of the influence function of locally asymptotically linear estimators
> - 10.1097/00001648-199203000-00013 may be a valid DOI for title: Identifiability and exchangeability for direct and indirect effects
> - 10.1007/978-1-4419-9782-1_27 may be a valid DOI for title: Cross-validated targeted minimum-loss-based estimation
> - 10.1257/aer.p20171038 may be a valid DOI for title: Double/debiased/neyman machine learning of treatment effects
> - 10.1007/978-3-319-59626-6_5 may be a valid DOI for title: The method of path coefficients
> - 10.4324/9780203994627 may be a valid DOI for title: The Logic of Scientific Discovery
> - 10.1097/01.ede.0000208475.99429.2d may be a valid DOI for title: Estimation of direct causal effects
>
> INVALID DOIs
>
> - None
> ```

@nhejazi Please fix these missing DOIs and let us know when you are ready for the review to start.
