### 2021-12-08 #3979 @nhejazi
thanks for flagging these initial revisions @mikldk. these have been addressed as of the most recent commit, https://github.com/nhejazi/medoutcon/commit/d0e65f5df58b5f5bf9917ad0a4766e343d50599d.
