### 2021-12-11 nhejazi/medoutcon#31 @erikcs
Running README example requires extra dependency
After I installed medoutcon from github, a big handful of dependencies, and dependencies' dependencies were updated/installed. When I run the readme example I get the following, which may seem a bit quirky for a new user:

```
> # compute targeted minimum loss estimate of the interventional direct effect
> tmle_de <- medoutcon(W = example_data[, ..w_names],
+                      A = example_data$A,
+                      Z = example_data$Z,
+                      M = example_data[, ..m_names],
+                      Y = example_data$Y,
+                      effect = "direct",
+                      estimator = "tmle")
Error in requirePackages(private$.required_packages, why = class(self)[1],  :
  For Lrnr_glm_fast please install the following packages: speedglm
```
