### 2021-12-11 nhejazi/medoutcon#31 @nhejazi
Thanks for catching this and pointing it out. This seems to happen because the default algorithms used for nuisance estimation wrap the `speedglm` package via `sl3` (https://github.com/nhejazi/medoutcon/blob/master/R/medoutcon.R#L81-L85), while `speedglm` is only listed in the `Suggests` of this package. This could be fixed by moving `speedglm` to `Depends`, I believe. The same is true of `hal9001` also (https://github.com/nhejazi/medoutcon/blob/master/R/medoutcon.R#L86-L87).
