### 2021-12-16 #3979 @nhejazi
@erikcs, thanks for your review --- I've noted the issue opened and will address it at the same time as I make other changes to the package for this submission.

@rrrlw, thanks for agreeing to review this package + paper --- I'll try integrate any changes you suggest as soon as your review is available and will resolve other comments alongside yours in a single pass.
