### 2021-12-20 nhejazi/medoutcon#32 @rrrlw
Minor issue: LICENSE
The MIT license is appropriate. However, you need a copy of the plaintext license (https://opensource.org/licenses/MIT) in the repository per JOSS guidelines.

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/3979).
