### 2021-12-20 nhejazi/medoutcon#32 @rrrlw
I now see the MIT license in the README - moving it to the LICENSE file or a new LICENSE.md file should be sufficient.
