### 2021-12-20 nhejazi/medoutcon#33 @rrrlw
Optional issue: examples
A comprehensive example is provided in the README. However, users of the R package would also benefit from quick access to examples in the documentation itself (i.e. using the `@example` or `@examples` Roxygen tag so that it can be accessed with `?<function name>`).

If adding an example within the documentation is too complex, you can simply close this issue.

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/3979).
