### 2021-12-20 #3979 @rrrlw
[Minor issue](https://github.com/nhejazi/medoutcon/issues/32) - after this and @erikcs's [issue](https://github.com/nhejazi/medoutcon/issues/31) are resolved, looks good to go. Thank you for this contribution, @nhejazi!
