### 2021-12-20 b990c07
add MIT license and update dates
