+++ README.Rmd
@@ -194,7 +194,7 @@ for details:
 ```
 MIT License
 
-Copyright (c) 2020-2021 Nima S. Hejazi
+Copyright (c) 2020-2022 Nima S. Hejazi
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
