### 2021-12-20 4d5d15a
minor version bump; tweak deps for #31
