+++ inst/REFERENCES.bib
@@ -22,9 +22,9 @@
   url = {https://arxiv.org/abs/1912.09936},
   doi = {10.1093/biomet/asaa085},
   journal={Biometrika},
-  volume={},
-  number={},
-  pages={},
+  volume = {108},
+  number = {3},
+  pages = {627--641},
   publisher={Oxford University Press}
 }
 
