### 2021-12-20 nhejazi/medoutcon#32 @nhejazi
Thanks for pointing this out, I'd neglected to include a `LICENSE.md` file but agree that it's a good practice. This has been addressed in b990c07 of #34.
