### 2021-12-20 nhejazi/medoutcon#31 @nhejazi
Following up on this to note that 4d5d15a in #34 moved both `hal9001` and `speedglm` to the `Depends` slot. In testing the `README` example in a clean R session (`R --vanilla`), the example appears to work without issue; that is, I cannot reproduce the error. If you could spare a minute to try with that package version and confirm, that'd be helpful. If not, I'll try to run on another machine just to double check. Thanks again for your review of this package @erikcs.
