### 2021-12-20 nhejazi/medoutcon#33 @nhejazi
Thanks for pointing this out. As someone who does access examples through the R documentation, I agree that it would be very useful to have these for the exported functions (of which there are only 3, `medoutcon`, `est_tml`, and `est_onestep`). I'll add these later today and update once that has been completed.
