### 2021-12-21 0980f97
remove export of internal estimation functions; add example to eponymous function
