### 2021-12-21 nhejazi/medoutcon#33 @nhejazi
In 0980f97 of #34, I removed the exporting of the internal estimation functions `est_tml()` and `est_onestep()`, and I added an example based on the `README` to the eponymous function `medoutcon()`.
