### 2021-12-21 nhejazi/medoutcon#31 @nhejazi
OK, I've now tried the example again on a completely fresh machine and believe that this issue has been resolved.
