### 2021-12-21 #3979 @nhejazi
The points of improvement brought up in issues 31, 32, and 33 have been addressed in PR 34. Thanks very much @erikcs and @rrrlw for your timely and diligent reviews!

@mikldk I’ve left PR 34 open in case you might have any final revision comments. If not (or if everything looks in order), I’ll go ahead and merge that PR.
