### 2021-12-21 #3979 @nhejazi
Great! Thanks for the quick confirmation @mikldk. I just merged the PR and can take other steps (e.g., minting a release DOI) as needed. I've only shortened the package title and would like to keep the paper title a bit more general, so I think that can stay as is. I did also introduce a minor version bump 0.1.5 -> 0.1.6, and I think the version of the package referenced in the opening comment of this issue should reflect that.
