### 2021-12-23 #3979 @mikldk
@rrrlw Can you confirm that the license issue has been resolved (and check the corresponding box at the top of this issue)?
