### 2021-12-29 f71d240
final touch-ups to JOSS ppaer
