+++ inst/REFERENCES.bib
@@ -134,10 +134,10 @@
   doi = {10.5281/zenodo.1342293},
   url = {https://github.com/tlverse/sl3},
   note = {{\texttt{R}} package},
-  version = {1.4.3},
+  version = {1.4.4},
   author = {Coyle, Jeremy R and Hejazi, Nima S and Malenica, Ivana and
     Phillips, Rachael V and Sofrygin, Oleg},
-  title = {{\texttt{sl3}}: Modern pipelines for machine learning and {Super
+  title = {{\texttt{sl3}}: Modern Machine Learning Pipelines for {Super
     Learning}},
   year={2021}
 }
@@ -291,7 +291,7 @@
 
 @article{coyle2018origami,
   author = {Coyle, Jeremy R and Hejazi, Nima S},
-  title = {origami: A Generalized Framework for Cross-Validation in R},
+  title = {{origami}: A Generalized Framework for Cross-Validation in R},
   journal = {Journal of Open Source Software},
   volume = {3},
   number = {21},
