### 2021-12-29 #3979 @nhejazi
Thanks, @mikldk -- I've completed a final pass through the paper and package documentation, making minor fixes throughout. The PDF created by whedon looks good to me, so I went ahead and created a GitHub release (https://github.com/nhejazi/medoutcon/releases/tag/joss) and archived that release via Zenodo (https://zenodo.org/record/5809520). The archived package is v0.1.6 with DOI 10.5281/zenodo.5809520.
