### 2022-01-05 #3979 @kthyng
Archive and version look good.

Paper comments:
* Coyle and Hejazi 2018 has a lowercase instead of uppercase "R"
* please check other capitalization in references.
