+++ inst/REFERENCES.bib
@@ -3,7 +3,7 @@
     interventional (in)direct effects},
   author = {Hejazi, Nima S and Rudolph, Kara E and {van der Laan}, Mark J and
     D{\'\i}az, Iv{\'a}n},
-  year = {2021},
+  year = {2022},
   url = {https://arxiv.org/abs/2009.06203},
   doi = {},
   journal = {\textnormal{revision invited at} Biostatistics},
@@ -67,8 +67,8 @@
 }
 
 @book{vdl2011targeted,
-  title={Targeted learning: causal inference for observational and experimental
-         data},
+  title={Targeted Learning: Causal Inference for Observational and Experimental
+    Data},
   author={{van der Laan}, Mark J and Rose, Sherri},
   year={2011},
   publisher={Springer Science \& Business Media}
@@ -86,7 +86,7 @@
   title={Longitudinal mediation analysis with time-varying mediators and
     exposures, with application to survival outcomes},
   author={Zheng, Wenjing and {van der Laan}, Mark J},
-  journal={Journal of causal inference},
+  journal={Journal of Causal Inference},
   volume={5},
   number={2},
   year={2017},
@@ -98,7 +98,7 @@
   title={Effect decomposition in the presence of an exposure-induced
     mediator-outcome confounder},
   author={VanderWeele, Tyler J and Vansteelandt, Stijn and Robins, James M},
-  journal={Epidemiology (Cambridge, Mass.)},
+  journal={Epidemiology},
   volume={25},
   number={2},
   pages={300},
@@ -123,7 +123,7 @@
 @article{vdl2007super,
   title={Super learner},
   author={{van der Laan}, Mark J and Polley, Eric C and Hubbard, Alan E},
-  journal={Statistical applications in genetics and molecular biology},
+  journal={Statistical Applications in Genetics and Molecular Biology},
   volume={6},
   number={1},
   year={2007},
@@ -156,7 +156,8 @@
 @incollection{zheng2011cross,
   title={Cross-validated targeted minimum-loss-based estimation},
   author={Zheng, Wenjing and {van der Laan}, Mark J},
-  booktitle={Targeted Learning},
+  booktitle={Targeted Learning: Causal Inference for Observational and
+    Experimental Data},
   pages={459--474},
   year={2011},
   doi={10.1007/978-1-4419-9782-1_27},
@@ -234,13 +235,11 @@
   author = {{Tchetgen Tchetgen}, Eric J and Shpitser, Ilya},
   doi = {10.1214/12-AOS990},
   journal = {Annals of Statistics},
-  language = {eng},
-  month = {jun},
-  number = {3},
-  pages = {1816--1845},
   title = {Semiparametric Theory for Causal Mediation Analysis: efficiency
     bounds, multiple  robustness, and sensitivity analysis},
   volume = {40},
+  number = {3},
+  pages = {1816--1845},
   year = {2012}
 }
 
@@ -277,21 +276,21 @@
   author={Bickel, Peter J and Klaassen, Chris AJ and Ritov, YA'Acov and
     Wellner, Jon A},
   year={1993},
-  publisher={Johns Hopkins University Press Baltimore}
+  publisher={Johns Hopkins University Press}
 }
 
 @manual{R,
   address = {Vienna, Austria},
   author = {{R Core Team}},
   organization = {R Foundation for Statistical Computing},
-  title = {\text{R}: A Language and Environment for Statistical Computing},
+  title = {{R}: A Language and Environment for Statistical Computing},
   url = {https://R-project.org/},
-  year = {2021}
+  year = {2022}
 }
 
 @article{coyle2018origami,
   author = {Coyle, Jeremy R and Hejazi, Nima S},
[…]
