### 2022-01-05 #3979 @nhejazi
thanks for going through the paper and recommending this extra pass @kthyng. I've fixed the typo you spotted as well as several others in https://github.com/nhejazi/medoutcon/commit/f1a419af7a31f1db29e0a8c672c4aa18b4630a78
