### 2023-10-06 #3979 @ajstewartlang
:wave: @nhejazi Might you perhaps have the bandwidth to help us out by reviewing the following submission please? I think you'll find it interesting as it seems to nicely overlap with your interests and expertise:

https://github.com/openjournals/joss-reviews/issues/5862
