### 2023-10-13 #3979 @nhejazi
hi @ajstewartlang i think i can find some time to review this submission. please do note, though, that i did overlap with the author in graduate school, so i leave it to you to make sure that this is ok with JOSS policies (we have not worked together)
